Added timeout logic for prometheus metrics
For xhttp_prom module there were added an optional timeout
values for every metric (counter, gauge, histogram)

Added API function 'metric_set_timeout' for xhttp_prom module
to be able to set this timeout value.

Timeout value is pointed in minutes. If timeout is set for a
metric and during the timeout time there are no metric changes,
then this metric is removed.

If timeout is not set for a metric then the global timeout
is applicable for this metric. If timeout is set for a metric
then the global timeout is not used for this particular metric.

JIRA: SBC-24